macos wip

+++ make2/cpbin.sh
@@ -25,11 +25,22 @@ cp ../bin/${jplatform}/j64/libj.so ../jlibrary/bin/. || true
 
 elif [ "`uname -m`" = "arm64" ]; then
 
+# bin for arm64 binaries
+# MUST rename/remove object first; overwrite cause cache error
+if [ -f "../bin/${jplatform}/j64arm/jconsole" ]; then
+mv ../jlibrary/bin/jconsole ../jlibrary/bin/jconsole.old || true
+mv ../jlibrary/bin/libtsdll.dylib ../jlibrary/bin/libtsdll.dylib.old || true
+mv ../jlibrary/bin/libj.dylib ../jlibrary/bin/libj.dylib.old || true
 cp ../bin/${jplatform}/j64arm/jconsole ../jlibrary/bin/. || true
 cp ../bin/${jplatform}/j64arm/libtsdll.dylib ../jlibrary/bin/. || true
 cp ../bin/${jplatform}/j64arm/libj.dylib ../jlibrary/bin/. || true
+fi
 
+# bin32 for x86_64 binaries
 if [ -f "../bin/${jplatform}/j64/jconsole" ]; then
+mv ../jlibrary/bin32/jconsole ../jlibrary/bin32/jconsole.old || true
+mv ../jlibrary/bin32/libtsdll.dylib ../jlibrary/bin32/libtsdll.dylib.old || true
+mv ../jlibrary/bin32/libj.dylib ../jlibrary/bin32/libj.dylib.old || true
 cp ../bin/${jplatform}/j64/jconsole ../jlibrary/bin32/. || true
 cp ../bin/${jplatform}/j64/libtsdll.dylib ../jlibrary/bin32/. || true
 cp ../bin/${jplatform}/j64/libj.dylib ../jlibrary/bin32/. || true
